+++ src/config/aiModels/index.ts
@@ -23,6 +23,7 @@ import { default as minimax } from './minimax';
 import { default as mistral } from './mistral';
 import { default as moonshot } from './moonshot';
 import { default as novita } from './novita';
+import { default as nvidia } from './nvidia';
 import { default as ollama } from './ollama';
 import { default as openai } from './openai';
 import { default as openrouter } from './openrouter';
@@ -37,6 +38,7 @@ import { default as taichu } from './taichu';
 import { default as tencentcloud } from './tencentcloud';
 import { default as togetherai } from './togetherai';
 import { default as upstage } from './upstage';
+import { default as vllm } from './vllm';
 import { default as wenxin } from './wenxin';
 import { default as xai } from './xai';
 import { default as zeroone } from './zeroone';
@@ -85,6 +87,7 @@ export const LOBE_DEFAULT_MODEL_LIST = buildDefaultModelList({
   mistral,
   moonshot,
   novita,
+  nvidia,
   ollama,
   openai,
   openrouter,
@@ -99,6 +102,7 @@ export const LOBE_DEFAULT_MODEL_LIST = buildDefaultModelList({
   tencentcloud,
   togetherai,
   upstage,
+  vllm,
   wenxin,
   xai,
   zeroone,
@@ -128,6 +132,7 @@ export { default as minimax } from './minimax';
 export { default as mistral } from './mistral';
 export { default as moonshot } from './moonshot';
 export { default as novita } from './novita';
+export { default as nvidia } from './nvidia';
 export { default as ollama } from './ollama';
 export { default as openai } from './openai';
 export { default as openrouter } from './openrouter';
@@ -142,6 +147,7 @@ export { default as taichu } from './taichu';
 export { default as tencentcloud } from './tencentcloud';
 export { default as togetherai } from './togetherai';
 export { default as upstage } from './upstage';
+export { default as vllm } from './vllm';
 export { default as wenxin } from './wenxin';
 export { default as xai } from './xai';
 export { default as zeroone } from './zeroone';